backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic-settings==2.7.0
boto3==1.36.0
apscheduler==3.10.4
httpx==0.28.1
aiofiles==24.1.0
cryptography>=44.0.0
Pillow>=11.0.0
pypdf>=5.0.0
reportlab>=4.0.0
aiosqlite>=0.20.0
